move monitoring file to prevent collision during integration test run
