+++ tests/integration/monitoring-workflow.js
@@ -1,7 +1,7 @@
 const BbPromise = require('bluebird')
 const path = require('path')
 
-const idioms = require('../idioms')
+const idioms = require('./idioms')
 
 const modifyScript = () => idioms.loadAndMerge(path.join(__dirname, 'script.yml'), {
   config: {
